move filters to seperate module
